Rename forkCounter.ts to gossipForkCounter.ts
