+++ ironfish-cli/src/utils/gossipForkCounter.ts
